Hata düzeltmeleri
